Add Hard Mode and EasyMotion Plugins
I am trying to avoid using arrow keys to navigate and even the hjkl keys. Using this hard mode plugin may make it harder to navigate if you don't know how, but things like 3G to jump to line 3 and f" to jump to a quote in sentence make it easier. Also, added the EasyMotion plug-in to make it even more easier. Finally, I defined a mapping to Toggle Hard Mode.